+++ cvat/apps/dataset_manager/bindings.py
@@ -949,7 +949,10 @@ class ProjectData(InstanceLabelData):
 
     def _init_tasks(self):
         self._db_tasks: OrderedDict[int, Task] = OrderedDict(
-            ((db_task.id, db_task) for db_task in self._db_project.tasks.order_by("subset","id").all())
+            (
+                (db_task.id, db_task)
+                for db_task in self._db_project.tasks.exclude(data=None).order_by("subset","id").all()
+            )
         )
 
         subsets = set()